[Refactor] Consolidate User model definition and update imports
- Moved the User model definition from `user.py` to `base.py`, streamlining the model structure and improving organization.
- Updated the import statement in `__init__.py` to reflect the new location of the User model, ensuring base models are imported first.
- This change enhances code clarity and maintains consistency in model management.
